refactor(backoffice-v2): addressed pr comments
