+++ apps/backoffice-v2/src/common/utils/hash-string/hash-string.ts
@@ -1,0 +1,10 @@
+export const hashString = (str: string) => {
+  let hash = 0;
+
+  for (let i = 0; i < str.length; i++) {
+    hash = (hash << 5) - hash + str.charCodeAt(i);
+    hash |= 0; // Convert to 32-bit integer
+  }
+
+  return hash;
+};
