+++ apps/backoffice-v2/src/common/utils/string-to-rgb/string-to-rgb.ts
@@ -1,13 +1,4 @@
-export const hashString = (str: string) => {
-  let hash = 0;
-
-  for (let i = 0; i < str.length; i++) {
-    hash = (hash << 5) - hash + str.charCodeAt(i);
-    hash |= 0; // Convert to 32-bit integer
-  }
-
-  return hash;
-};
+import { hashString } from '@/common/utils/hash-string/hash-string';
 
 export const stringToRGB = (str: string) => {
   // Generate a hash of the string
