+++ apps/backoffice-v2/src/pages/Statistics/Statistics.page.tsx
@@ -1,16 +1,16 @@
 import React, { FunctionComponent } from 'react';
-import { UserStats } from '@/pages/Statistics/components/UserStats/UserStats';
-import { PortfolioRiskStats } from '@/pages/Statistics/components/PortfolioRiskStats/PortfolioRiskStats';
-import { WorkflowStats } from '@/pages/Statistics/components/WorkflowStats/WorkflowStats';
+import { UserStatistics } from '@/pages/Statistics/components/UserStatistics/UserStatistics';
+import { PortfolioRiskStatistics } from '@/pages/Statistics/components/PortfolioRiskStatistics/PortfolioRiskStatistics';
+import { WorkflowStatistics } from '@/pages/Statistics/components/WorkflowStatistics/WorkflowStatistics';
 
 export const Statistics: FunctionComponent = () => {
   return (
     <div>
       <h1 className={'pb-5 text-2xl font-bold'}>Statistics</h1>
       <div className={'flex flex-col space-y-8'}>
-        <UserStats fullName={'John Doe'} />
-        <PortfolioRiskStats />
-        <WorkflowStats />
+        <UserStatistics fullName={'John Doe'} />
+        <PortfolioRiskStatistics />
+        <WorkflowStatistics />
       </div>
     </div>
   );
