+++ apps/backoffice-v2/src/pages/Statistics/components/PortfolioRiskStatistics/constants.ts
@@ -1,0 +1,13 @@
+export const riskLevelToFillColor = {
+  low: 'fill-success',
+  medium: 'fill-warning',
+  high: 'fill-destructive',
+  critical: 'fill-foreground',
+} as const;
+
+export const riskLevelToBackgroundColor = {
+  low: 'bg-success',
+  medium: 'bg-warning',
+  high: 'bg-destructive',
+  critical: 'bg-foreground',
+} as const;
