+++ apps/kyb-app/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@ballerine/kyb-app",
   "private": true,
-  "version": "0.3.14",
+  "version": "0.3.15",
   "type": "module",
   "scripts": {
     "dev": "vite",
@@ -14,10 +14,10 @@
     "test:dev": "vitest"
   },
   "dependencies": {
-    "@ballerine/common": "^0.9.8",
+    "@ballerine/common": "^0.9.9",
     "@ballerine/blocks": "0.2.3",
     "@ballerine/ui": "0.5.3",
-    "@ballerine/workflow-browser-sdk": "0.6.14",
+    "@ballerine/workflow-browser-sdk": "0.6.15",
     "@lukemorales/query-key-factory": "^1.0.3",
     "@radix-ui/react-icons": "^1.3.0",
     "@rjsf/core": "^5.9.0",
