+++ examples/headless-example/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ballerine/headless-example
 
+## 0.3.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.9
+  - @ballerine/workflow-browser-sdk@0.6.15
+
 ## 0.3.13
 
 ### Patch Changes
