+++ packages/common/src/index.ts
@@ -31,7 +31,7 @@ export type {
   TDefaultSchemaDocumentPage,
   TDocument,
 } from './schemas';
-export type { AnyRecord, LoggerInterface, Serializable } from './types';
+export type { AnyRecord, LoggerInterface, Serializable, SortDirection } from './types';
 
 export {
   WorkflowDefinitionConfigThemeSchema,
