+++ packages/workflow-core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ballerine/workflow-core
 
+## 0.6.15
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.9
+
 ## 0.6.14
 
 ### Patch Changes
