+++ sdks/web-ui-sdk/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web-ui-sdk
 
+## 1.5.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.9
+
 ## 1.5.9
 
 ### Patch Changes
