+++ sdks/workflow-browser-sdk/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ballerine/workflow-browser-sdk
 
+## 0.6.15
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.9
+  - @ballerine/workflow-core@0.6.15
+
 ## 0.6.14
 
 ### Patch Changes
