+++ sdks/workflow-browser-sdk/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@ballerine/workflow-browser-sdk",
   "author": "Ballerine <dev@ballerine.com>",
-  "version": "0.6.14",
+  "version": "0.6.15",
   "description": "workflow-browser-sdk",
   "module": "./dist/esm/index.js",
   "main": "./dist/cjs/index.js",
@@ -33,8 +33,8 @@
     "node": ">=12"
   },
   "dependencies": {
-    "@ballerine/common": "0.9.8",
-    "@ballerine/workflow-core": "0.6.14",
+    "@ballerine/common": "0.9.9",
+    "@ballerine/workflow-core": "0.6.15",
     "xstate": "^4.37.0"
   },
   "devDependencies": {
